Huge commit.
- Changed all the long type time variables to int.
- Rewrite and enhance the LyricsProcessor with clearer code and fully utilize it in Viewport.
- Rewrite Viewport class a lot and add a LyricsArea class to help make cache areas for same line display.
- Remove all references of SaveLoadManager.textList.
- Rearrange code in classes.

+++ src/main/java/net/okt/system/command/marks/MarkAddCommand.java
@@ -5,10 +5,10 @@ import net.okt.system.command.Command;
 import java.util.List;
 
 public class MarkAddCommand implements Command {
-    private final List<Long> markList;
-    private final long time;
+    private final List<Integer> markList;
+    private final int time;
 
-    public MarkAddCommand(List<Long> markList, long time) {
+    public MarkAddCommand(List<Integer> markList, int time) {
         this.markList = markList;
         this.time = time;
     }
